+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to the "my-vscode-hybris-tools" extension will be documented
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 
+
+- Init GroovyFS (Virtual FS)
+
 ## 1.0.10
 
 - Sql  - Support formatting